asus-nb-wmi: Ignore ScreenPad in Zenbook Duo keyboard quirk
For some reason, the Zenbook Duo UX8406MA has ScreenPad WMI methods in its
ACPI tables. It does not have a ScreenPad, and it's a fairly safe bet that
no future products in this line will either.

The ACPI tables do seem to be performing an EC read/write on the ScreenPad
brightness endpoint, but this has no visible effect on the device.

Stop exposing this non-existent ScreenPad to userspace.

Signed-off-by: Joshua Leivenzon <hacker1024@users.sourceforge.net>